backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.119.0",
    "google-auth>=2.41.1",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-aiplatform>=1.121.0",
    "google-genai>=1.45.0",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "langchain-openai>=0.3.35",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.37.0",
]
